Задание 4, 5,6

+++ src/Operators/Task5.java
@@ -3,14 +3,14 @@ package Operators;
 import java.util.Scanner;
 public class Task5 {
     public static void main(String[] args) {
-        int num1;
-        int num2;
-        int ans;
+        double num1;
+        double num2;
+        double ans;
         char op;
         Scanner reader = new Scanner(System.in);
         System.out.print("Введите 2 числа: ");
-        num1 = reader.nextInt();
-        num2 = reader.nextInt();
+        num1 = reader.nextDouble();
+        num2 = reader.nextDouble();
         System.out.print("Выберете один из символов ‘+’, ‘-’, ‘*’ или ‘/’.: ");
         op = reader.next().charAt(0);
         switch(op) {
